Store session in password to support BitlBee
Account settings in Purple don't get saved in BitlBee, which means
BitlBee users need to reauthenticate every time they log in.

Purple does support changing the password in BitlBee using a set
password signal. Change session storage so that all keys are stored as a
space-separated string in the password field that can be persisted in
BitlBee.