Ensure buddies exist when message received
If a message arrives from and the user is pulling contacts, then add the
buddy to the contact list if they're not there already. Sometimes the
messages come before the buddy list.

In Bitlbee, a message from someone not in your buddy list gets sent to
the control channel. If messages arrive before the buddy information,
the control channel gets littered. This fix prevents that.