Create group chats for group chats
Instead of group chats being represented as "virtual buddies", represent
them as actual group chats.

The "virtual buddy" approach does not work in Bitlbee as it will ignore
messages not from the buddy of the conversation. By using group chats we
can add all messaging users to it so the messages are displayed.

Implement a bunch of the Pidgin infrastructure around this, including
listing available group chats so that a Bitlbee user can initiate a
conversation (group chats in the buddy list are not displayed by
Bitlbee).